fix: refine agent fast path batching
